bazel/linux: test cleanups, and qemu_test target (#635)
* bazel/linux: remove compatibility rules.

Users of those rules have been updated to point to the new location,
no reason to keep variables around to provide access to the old rules.

* bazel/linux: rename kernel_test to kunit_test, move in test.bzl file.

* bazel/linux: add a qemu_test target.

Background:
So far, we've had a qemu_run target. This only allows to run a target,
but does not tell bazel that the target is actually a test.

In this PR:
- add a small qemu_test macro, that when used instantiates a qemu_run
  target, and an exec_test() target.

This allows to run qemu with full access to the terminal, as well as run
it as part of an automated test suite.

So: qemu_run -> defines a target to run manually to land in qemu.
    qemu_test -> defines a qemu_run taget, + an automated test target.

Additionally:
- minor fix in the kunit_test rule, where kwargs were not propagated
  correctly to the exec_test target (possibly affecting visibility
  and similar).

Tested:
Small change in showcase PR seems to make it work correctly.

This should help with SF-162 ans SF-150.