bazel/linux: propagate the exit status outside the emulator
Success or failure inside the emulator is propagated to the exit status
of bazel run. The init script that executes in the emulator stores its
exit status in a temporary file. The script that executes the emulator
adjusts its own exit status based on this file.

Jira: https://enfabrica.atlassian.net/browse/SF-165
Signed-off-by: George Prekas <george@enfabrica.net>