Libraries to write our own CODEOWNERS/OWNERS bot.
Specifically, this PR includes:
- an initial library to parse OWNERS/CODEOWNERS files as well as protocol
  buffers in our repository as owners files, and some tests.

The thinking here is that we'd have backward compatibility with the existing
CODEOWNERS file, and some simple syntax to support OWNERS file, as well as
supporting a basic protocol buffer that allows full flexibility.

The library - albeit "code complete/usable" - is unfinished, as I have not
written the remaining tests, and there are some known bugs/missing features.