astore_package: Raise sha256sum timeout
This change raises the timeout for calculating the sha256 of items
downloaded from enkit. This timeout is currently causing false failures
in presubmits - while an individual sha256 should be relatively fast
compared to the old timeout, it is likely that the Cloud Build machine
is downloading/unpacking other items concurrently and is therefore
slower than expected.

Jira: INFRA-1179