enproxy: Use negative port to signal auto port detection
It's useful to create a mechanism by which the proxy can deduce the
correct target port for `enkit tunnel` based on the host - the first
usecase is by doing SRV record lookup for Consul DNS names.

This change modifies `enkit tunnel` to specify a negative port number to
enproxy when no port is provided on the commandline. Enproxy interprets
a negative port as requesting port resolution.

Because port can now be negative, the type of port must change from
uint16 to int everywhere.

Jira: INFRA-1223