kunit: append the kunit command steps to the runner steps
This patch swaps the order of the 'runner_cfg' and 'cfg' parameters in
the mcreate_rule() call for kunit tests.  The net effect is the
commands from 'cfg' are appended to the 'runner_cfg' commands, instead
of the other way around.

We want the 'runner_cfg' commands to run first, as their job is to
setup the test environment for the kunit test.  We do not want to run
the kunit test and then setup the environment.

Signed-off-by: Curt Brune <curt@enfabrica.net>