kunit_test: default to QEMU runner
This patch changes the default runner for the kunit_test() rule from
the UML runner to the QEMU runner.

Also, this patch adds a kernel argument, 'printk.time=0', which
disables the printing of timestamps for printk message.  Without this
flag, the kunit parser is unable to parse the test results.

Signed-off-by: Curt Brune <curt@enfabrica.net>