kunit: updates for kunit inside of QEMU (#667)
* kunit: do mot mount sysfs on /sys automatically

This patch removes the automatic mount of sysfs on /sys for a couple
of reasons:

1. the kunit framework now uses 'insmod' instead of 'modprobe', so the
original reason for mounting sysfs no longer exists.

2. kunit_test() can be run with a custom runner that can handle any
needed "pre-test" steps, like mounting sysfs or procfs, etc.

For the case of using a QEMU runner, we will use a common system setup
set of run commands that already handles mounting various kernel file
systems.

Relates-to: MCU-85

Signed-off-by: Curt Brune <curt@enfabrica.net>

* kunit: append the kunit command steps to the runner steps

This patch swaps the order of the 'runner_cfg' and 'cfg' parameters in
the mcreate_rule() call for kunit tests.  The net effect is the
commands from 'cfg' are appended to the 'runner_cfg' commands, instead
of the other way around.

We want the 'runner_cfg' commands to run first, as their job is to
setup the test environment for the kunit test.  We do not want to run
the kunit test and then setup the environment.

Signed-off-by: Curt Brune <curt@enfabrica.net>

* kunit_test: default to QEMU runner

This patch changes the default runner for the kunit_test() rule from
the UML runner to the QEMU runner.

Also, this patch adds a kernel argument, 'printk.time=0', which
disables the printing of timestamps for printk message.  Without this
flag, the kunit parser is unable to parse the test results.

Signed-off-by: Curt Brune <curt@enfabrica.net>

* kernel_test: use kunit_test() with a UML runner

Previously, kernel_test was a straight alias for kunit_test.  The
current consumers of kernel_test assume the runner is the UML runner,
however kunit_test now uses the QEMU runner by default.

This patch rewrites kernel_test as a wrapper around kunit_test passing
in a UML runner.  This allows legacy kernel_test consumer to continue
to work until they are converted to use the kunit_test directly.

Signed-off-by: Curt Brune <curt@enfabrica.net>