gee: add -m option to gcd (#645)
gee: add -m option to gcd

As a convenience option, I added a "-m" option to the gcd command, that
causes gcd to create a new branch that is a child of the master branch.

The "-b" option functionality (creation of a new branch that is a child
of the current branch) remains unchanged.

I also updated "gee rmbr" to be able to operate on a list of branches instead
of a single branch, and removed the "booby trap" where running "gee rmbr"
without any options removes the current branch.

This PR also improves gee's documentation a little bit.

* The "-m" option creates a new branch that is a branch of master

Tested: use "gcd -b" and "gcd -m" to create new branches, and used
"gee getparent" to confirm that the corrent branch lineage was created.
I then used "gee rmbr" to remove all of the newly created test branches
using a single command.