bazel/utils: initial set of rules to run an arbitrary target on a remote machine. (#677)
* bazel/utils: initial set of rules to run an arbitrary target on a remote machine.

See the top of remote.bzl file for an explanation of the ruleset added.

Why? Typical development cycle has developers build scripts or tests
locally, and wanting to try them on a remote machine. I expect we'll have
something better and more dynamic with machine allocations in the future,
but this is a starting point for manual testing.

SF-218 for background as well.

This is a first step: qemu run rules will be modified to better play
with the remote rules, and allow baremetal run, and we'll need an
official wrapper to prepare the environment on the remote machine.

* Add propagation of environment variables.