gee prompt: fix labeling of non-printing characters (#696)
gee bash_setup: fix non-printing characters in bash

I got annoyed that the gee-configured prompt wasn't perfect so I fixed it.

In bash: termcap and readline have a weird interaction where readline
doesn't understand which characters are non-printing and which aren't.
Bash has a prompting-specific workaround for this problem, which is to
explicitly label non-printing characters within "\[" and "\]" strings.

This PR correctly adds those labels so that readline will not corrupt
the column index when interactively editing a commandline.

Tested: manually.