third_party: Create dummy go packages for import paths
This change creates a few dummy Go packages to satisfy native Go tooling
such as `go get`.

The affected dir contains code copied from other repositories via
Copybara. The copied code had its directory structure preserved to
minimize build issues, but the importpath is shortened and doesn't match
the path in the repository out of convenience (the paths are awfully
long), breaking convention.

This change adds a dummy Go package to the shortened import path, so
that when go tooling encounters the shortened path in an import
statement, it can find the package under the dummy path and not error.

The dummy package dirs are excluded from the Copybara config, so that
the next Copybara import doesn't inadvertently delete them.

Tested:
* All build files are gazelle-clean
* `go get -u ./...` no longer complains about packages in the affected
  dirs