nomad: Add event listener
This change adds a binary that listens to Nomad via its event stream
API, and exposes counts of events via metrics.

The impetus for this binary is to provide a more reliable counter of OOM
events, but it's possible we will want to count other events as well.

Tested: Manually, against actual Nomad installation with alloc that
reliably OOMs.