bazel/astore/defs.bzl: allow astore_download a longer timeout (#684)
bazel/astore/defs.bzl: allow astore_download a longer timeout

Another user experienced a presubmit failure that looks like bazel
timed out and terminated an "astore download" run action.  I don't
know how long bazel's default timeout is, but there seems to be
an undocumented way to change that timeout for run actions.

This PR attempts to set that timeout to 10 minutes, under the
philosophy that it's better for a build action to be slow than
to fail.

While testing this code, I learned that the DefaultInfo returned
by `astore_download` was incorrect in so far as the downloaded
result could not be used as an input for another rule, because
it was missing from the runfiles attribute.  Fixed.

Tested: added a quick test that downloads a file and compares
a checksum of the file against a precomputed value.