enkit outputs: Take local tunnel port as flag
Currently, only one user can `enkit outputs mount` on a shared dev
machine, as when tunnels are required the local port used is hardcoded
to a single value.

This change moves the value to a flag, so that multiple people might be
able to use `enkit outputs mount` by modifying the flag value.

The flag is defaulted to the same value as the hardcoded value, to keep
this change non-breaking.

Jira: INFRA-1476