gee gcd: fix incorrect mapping of branch-to-dir (#713)
gee gcd: fix incorrect mapping of branch-to-dir

@hanu-enf reported a bug in which "gcd master" was taking him to a feature
branch directory instead of master.  We root-caused the problem to gee's
parsing of the output of `git worktree list --porcelain` which produces
output like this if a branch is in a rebase-in-progress state:

```
worktree /home/hanu/gee/internal/master
HEAD b144e5c34ff7bd8f24717337bde4bfd03fc649d0
branch refs/heads/master

worktree /home/hanu/gee/internal/cmn_reg
HEAD b144e5c34ff7bd8f24717337bde4bfd03fc649d0
detached
```

`gee` was not parsing the `detached` line correctly.  This PR fixes the issue
by parsing the `detached` line and then falls back on a secondary technique to
find the branch name if the branch is in a rebase-in-progress state.

If all else fails, gee gives up and uses the basename of the worktree directory
as the branchname, which is almost always correct, but we prefer to use the
.git metadata as our ground truth whenever possible.

Tested: I put a branch into a rebase-in-progress state.  I observed the
failure.  Implemented workaround.  Ran `gee gcd` again for a variety of
directories and got the correct behavior.