bazel/linux: print qemu_test output to stdout (#720)
The output of commands executed by qemu_test is captured to a variable
and printed if the command fails. This behavior works fine with bazel
test which hides output by default. When the user executes a qemu_test
with bazel run, the expectation is to see the output in real time.

Signed-off-by: George Prekas <george@enfabrica.net>