fix gh-resolved (#719)
gee repair: fix incorrect gh-resolved settings

I'm not sure how, but for some users gh is configuring "origin" isntead
of "upstream" to be the base repo for managing PRs.  I still need to
root-cause why that happened to at least one user, but in the meanwhile
it's trivial to add a check to the "gee repair" command to detect the
condition and fix it.

I'm also changing the enkit agent environment variable check from looking
for SSH_AGENT_PID to SSH_AUTH_SOCK -- because in certain ssh-forwarded
environments, the latter will be set correctly even if the former isn't.

Tested: I manually broke my local git configuration, and then had gee
repair it.  I further verified that the gee repair command is harmless
in a healthy gee repo.