bazel affected-targets: Fix target name bug
The external-target-attribute-rewrite logic has a bug where it captures
a loop variable by reference, causing targets to get renamed with the
incorrect name. This change makes a copy of the loop variable and
assigns a reference to the copy instead, so that it is not mutated as
the loop continues.

Tested: `bazel affected-targets list --loglevel-console=info --start=4b551678e --end=swc_copy`
with this change doesn't yield erroneous targets, whereas it used to
roughly ~20% of the time.

Jira: INFRA-1548