gee gcd: redirect savelog output to stderr (#728)
gee gcd: redirect savelog output to stderr

Just a quick bugfix: gee periodically invokes `savelog` to rotate it's log
file.  `savelog` sometimes writes to stdout, which then interferes with
the `gee gcd` command, which requires that the only thing written to
stdout is supposed to be the generated path.

This PR quick-fixes the behavior so that savelog triggering a log rotation
won't interfere with the functioning of gcd.

Tested: inflated ~/gee.log to be larger than 64K, forcing a log rotation
on the next gee invocation, and then ran `gee gcd master > /tmp/foo.txt`
to capture the stdout of gee under those conditions.  Confirmed that
savelog output no longer pollutes our results.