gee: fix enkit auth invocation (#731)
gee: fix enkit auth invocation

While it appears that `enkit auth` is capable of inferring the username
to use during authentication, it turns out that it just caches this
information after the user does it for the first time.  We want gee
to work correctly on a completely clean install, and so the safest
option is to always specify the full `enkit auth $USER@enfabrica.net`
commandline.

Tested: rm ~/.config/enkit/* and then run `gee init`