turn on rerere for everyone all the time (#738)
gee: turn on rerere for all users

https://git-scm.com/docs/git-rerere

rerere is a feature that records the manual resolution of a conflict.  The
next time git runs into an identical conflict, it resolves it the same way.

This is useful when rebasing multiple branches.  For branches master->A->B,
if git rebase ends up renaming a commit when integrating from master into A,
that renamed commit becomes a merge conflict when merging from A into B.
With rerere, the manual resolution step when merging from master into A
is likely to be re-used when merging from A into B.

Tested: Still testing.