machinist server: Add HTTP endpoint for discovering nodes to scrape
Prometheus is currently fetching the list of nodes to scrape for metrics
from `mserver` via DNS A records. This means that metrics have
`instance=<IP>` on each, with no hope of ever mapping that back to a
hostname.

As it turns out, we can (and do) expose other types of DNS records, but
there is no way to get Prometheus both IP and hostname information via
DNS (it doesn't allow for fetching from the right kind of record, or
limits the fields it fetches from records that could contain both).

This change adds a separate mechanism that Prometheus does support: an
HTTP handler. By fetching `/metrics_targets`, a JSON array will be
generated that Prometheus understands that specifies scrape IPs and
their associated hostname. Port is not part of the address exposed, so
this must match node_exporter config and Prometheus config - but this is
no different than what we have today (A records don't export metrics
ports either).

Tested:
* Started server with `bazel run //machinist/mserver/cmd -- --loglevel-console=info`
* Fetch: `curl localhost:6433/metrics_targets` returns:

  ```
  []
  ```

* Started client with `bazel build //machinist/cmd && sudo bazel-bin/machinist/cmd/cmd_/cmd node poll --loglevel-console=info --auth-server=https://auth.corp.enfabrica.net --control-port=6433 --ips=192.168.1.220`
* Fetch: `curl localhost:6433/metrics_targets` returns:

  ```
  [{"labels":{"hostname":"pop-os"},"targets":["192.168.1.220"]}]
  ```