enkit tunnel: Support listening on Unix Domain Sockets
This change extends the functionality of the `-L`/`--listen` flag on
`enkit tunnel` to support Unix Domain Socket URLs as listen targets, in
the form `unix:///path/to/socket`. This involves:

* creating a function to contain the logic that normalizes the contents
  of the flag. The flag needs to continue to support `host:port`, as
  well as just a naked `port`, in addition to the new UDS path.
* fixing dependent code to not assume `TCPAddr`/`TCPConn`. In some
  places, type assertions are necessary, as TCP and Unix variants
  contain necessary methods not in the top-level interface; these should
  be relatively safe as long as we don't expand the network types to
  types that lack these methods.

Additionally, cleanup logic is improved by creating a Context that
becomes done on SIGINT, and plumbing that Context to relevant locations.
Care is taken to ensure that the listener is cleaned up at the
appropriate time depending on the listener's type and whether the tunnel
is running in background mode.

Tested:

Jira: