proxy/enproxy: improve enproxy_test to move more data and print status.
Before this PR:
enproxy_test end to end test moved 10Mb of data and passed if no
error was logged during the transfer.

In this PR:
- move 1GB of data instead.
- print progress as the transfer advances.

As is, the test now does not pass: most transfers will lock and
stop somewhere between 100 MB and 1 GB.

The PR adds progress logging, to ease seeing those problems as
the test is running.