proxy/tunnel: fix retry strategy.
This PR fixes a subtle bug introduced by `:=` and scoping if,

    if err := ...; err != nil {
      ...
    }

that caused a function to always succeed, even in case of error,
pretty much killing the reliability/retry strategy implemented
in the client, and allowing any simple error to freeze the connection.

PROBLEM DESCRIPTION:

Before this PR, the if was inverted, eg, "if the error is a close request,
return nil", otherwise return err:

    if err := ...; IsClose(err) {
        return nil
    }
    return err

except this last `return err`... would return the previous value of
err, nil! So this function would ALWAYS SUCCEED, disabling the retry
strategy.

PROBLEM FIX:

This PR inverts the if, simply.