proxy/tunnel: fix handling of short reads.
Background:
If an endpoint of a TCP connection is faster, a read() call may
return less data than expected (because it's not been sent yet,
it's in transit on the network, or ...).

The NASSH protocol expects to read 4 bytes indicating how long
the next chunk is. Those bytes are read from a websocket, which
I had originally assumed to be buffered/chunked, on top of http/http2,
and had expected 4 bytes to always be available.

Turns out that short reads/short buffers are much more likely when
transferring large amount of data over the internet.

The bug was known: a similar PR fixed the issue in the server a
few months ago, but the retry strategy papering over connection
issues and reconnects made me think it was a non issue.

In this PR:
- use the same strategy as the server to actually handle
  short reads cleanly.