fixing tunnel (#746)
* proxy/enproxy: improve enproxy_test to move more data and print status.

Before this PR:
enproxy_test end to end test moved 10Mb of data and passed if no
error was logged during the transfer.

In this PR:
- move 1GB of data instead.
- print progress as the transfer advances.

As is, the test now does not pass: most transfers will lock and
stop somewhere between 100 MB and 1 GB.

The PR adds progress logging, to ease seeing those problems as
the test is running.

* proxy/tunnel: fix retry strategy.

This PR fixes a subtle bug introduced by `:=` and scoping if,

    if err := ...; err != nil {
      ...
    }

that caused a function to always succeed, even in case of error,
pretty much killing the reliability/retry strategy implemented
in the client, and allowing any simple error to freeze the connection.

PROBLEM DESCRIPTION:

Before this PR, the if was inverted, eg, "if the error is a close request,
return nil", otherwise return err:

    if err := ...; IsClose(err) {
        return nil
    }
    return err

except this last `return err`... would return the previous value of
err, nil! So this function would ALWAYS SUCCEED, disabling the retry
strategy.

PROBLEM FIX:

This PR inverts the if, simply.

* proxy/tunnel: fix handling of short reads.

Background:
If an endpoint of a TCP connection is faster, a read() call may
return less data than expected (because it's not been sent yet,
it's in transit on the network, or ...).

The NASSH protocol expects to read 4 bytes indicating how long
the next chunk is. Those bytes are read from a websocket, which
I had originally assumed to be buffered/chunked, on top of http/http2,
and had expected 4 bytes to always be available.

Turns out that short reads/short buffers are much more likely when
transferring large amount of data over the internet.

The bug was known: a similar PR fixed the issue in the server a
few months ago, but the retry strategy papering over connection
issues and reconnects made me think it was a non issue.

In this PR:
- use the same strategy as the server to actually handle
  short reads cleanly.