enkit: add `version has-feature` subcommand
This change allows for a list of arbitrary "feature" strings to be
embedded into the enkit binary. These features could denote a new
capability within the binary, or a fix for a particular bug, and are
manually added by maintainers as fixes/features are added.

This change also adds a subcommand `has-feature` to `enkit version`.
This subcommand takes a list of feature strings, and exits 0 if they are
all present in this binary, or non-zero if at least one is missing.

Tested:
* `bazel run //enkit -- version has-feature tunnel-local-uds` - exits 0
* `bazel run //enkit -- version has-feature tunnel-local-uds foo` -
  exits nonzero