enkit: Enable cgo resolver for deployed linux binary
This change ensures that cgo is enabled and used in the Linux binary, so
that it can resolve addresses in the Cadence VCAD chamber machines,
where NIS is used (via nsswitch.conf/glibc).

It would seem that as soon as either goos/goarch and/or static is set on
the binary, rules_go uses a bazel "transition" that loses any `cgo`
attribute set on the binary. This is visible when building with
`--subcommands`, as the `CGO_ENABLED` env var is not set, and the
resulting binary cannot use the cgo DNS resolver.

Setting `pure=off` and not setting `static` seems to be the most
reliable way to get a working cgo binary.

Tested:
* Built with `bazel build //enkit:deploy`, which builds the versions
  that will actually be pushed to astore
* Ran `GODEBUG=netdns=cgo+2
  bazel-bin/enkit/deploy.sh.runfiles/enkit/enkit/enkit-linux-amd64_/enkit-linux-amd64
  ssh scott@build-04.sea` - saw that cgo resolver is actually used