enkit tunnel: Delete orphaned UDS sockets on start (#755)
Prior to this change, a tunnel could not be opened on a particular UDS
path if the socket already existed, with the assumption being that there
is another process potentially using that socket. Though the logic would
check for this by dialing the socket and looking for failure, it would
still error if the process controlling the previous socket instance had
died.

With this change, the socket is aggressively removed so that the new
instance can listen at that same path. The assumption is that it is rare
that the old process still exists if the socket dial fails (it is more
common that it was killed not-nicely, in which case removing the socket
is the correct fix).

Tested: with a change to using UDS for buildbarn in internal:
* `BAZEL_PROFILE=buildbarn bazel test //guidelines/...` - creates tunnels
* `BAZEL_PROFILE=buildbarn bazel test //guidelines/...` - reuses tunnels
* `ps -aux | grep tunnel` and `kill -9` on all PIDs
* `BAZEL_PROFILE=buildbarn bazel test //guidelines/...` - creates tunnels, no error

Jira: INFRA-1505