lib/bazel: set output_base from current client when doing things serially.
Background:
When doing operations serially, there is no reason to use a different
output_base (well..**). Using the current one will provide better caching
and locality, as it won't need to download and unpack everything from
scratch.

** only reason would be to "start fresh", and try things without
any sort of caching. But when run as a command line tool from a client
by a user, it should behave like bazel: re-use all the caching available.
If run from CI/CD, we have other knobs to control the cache.

In this PR:
- if no flags are specified to use a different behavior, default
  to using the current output_base when computing diffs serially.