kcerts: Fix unit tests
For some reason, the tests for `kcerts` used to be in their own package,
`kcerts_test`. A recent change modified the package to `kcerts` for one
of the test files, but not the other, confusing gazelle on rerun.

This change resolves the issue by making the unit tests part of the same
package (as is usually done) in all test files and re-running gazelle.

Tested: `bazel test //lib/...` passes