lib/bazel: add support for serial builds, more flags, more workspace (#757)
* enkit/lib/bazel: move the code to compute affected targets in a function.

This is in preparation for the next commit, where a serial strategy
will be introduced to re-use the output_base.

* /lib/bazel: add support for serial query or parallel query.

In this PR:
introduce the ability to run the affected-targets code
serially, rather than in parallel.

* lib/bazel: use an object to return from Serial or Parallel Query.

This is a stylistic cleanup: rather than return 5 values from each
function, just return an object with the 4 main values, and error.

* lib/bazel: allow to override output_base and bazel query from command line.

* lib/bazel: remove unused method.

* lib/bazel: minor workspace.go refactor to lazily load variables.

Background:
OpenWorkspace() runs a bunch of bazel commands at workspace
creation time, and carries a bunch of options along.

In a follow up commit I need:
1) ability to extract more information from bazel info.
2) ability to attach more information to the bazel commands.

In this PR:
- minor refactoring to turn Workspace() into a "lazy loader/evaluator",
  so that commands are actually run when needed.

* lib/bazel: set output_base from current client when doing things serially.

Background:
When doing operations serially, there is no reason to use a different
output_base (well..**). Using the current one will provide better caching
and locality, as it won't need to download and unpack everything from
scratch.

** only reason would be to "start fresh", and try things without
any sort of caching. But when run as a command line tool from a client
by a user, it should behave like bazel: re-use all the caching available.
If run from CI/CD, we have other knobs to control the cache.

In this PR:
- if no flags are specified to use a different behavior, default
  to using the current output_base when computing diffs serially.

* lib/bazel: add ability to pass extra arbitrary flags to bazel.

Before this commit: there was no way to pass arbitrary flags to
bazel.

After this commit: a command line option was added to allow
arbitrary flags to be passed to bazel.

* lib/bazel: improve debug output.

Specifically: print the exact bazel commands run in debug mode,
and print them in case of errors.

* lib/bazel: move cargo env out of exec library.

Background:
When running bazel in parallel, we need different cargo directories
to avoid one overwriting the other. This was implemented in the
exec function, setting a different CARGO_HOME for each invocation.

In this commit:
- add ability for callers to specify extra environment variables.
- move the logic up one level, and out of exec, so the caller
  can set it.
- make it so only the parallel runner sets it.
- improve debug output to include env variables, as they are
  clearly relevant.