astore: Add repository rule for downloading single file (#767)
We currently have an `astore_download` BUILD rule, which is a tad cursed
because network I/O shouldn't be done during the build phase (only
during the preceding fetch phase). In practice, this rule is problematic
to run on buildbarn, which doesn't have credentials to execute the
required fetch.

This change creates an equivalent `astore_file` rule, to complement the
existing rules (which assume the file is a tarball, and have some
unpacking logic). This rule covers the usecase where the file is not a
tarball, and saves us from adding release steps that involve wrapping
single files into a tarball.

Some of the common downloading logic is extracted to a helper function,
which helps shorten the definition of the new rule, as well as existing
rules.

Tested:
* added to `WORKSPACE`:

  ```
  load("//bazel/astore:defs.bzl", "astore_file")

  astore_file(
      name = "net_enfabrica_binary_enkit",
      digest = "b0731804c9ab398494f0457870087a817cb2a3c85f67b8ae0436f309b02066e0",
      executable = True,
      path = "tools/enkit",
      uid = "zz4jhhmn5e7m63f24tcppf7m77b62873",
  )
  ````

  `bazel build @net_enfabrica_binary_enkit//...` was successful

Jira: INFRA-1822