bazel/linux/runner: add support for hard coding wrapper flags. (#770)
Without this PR, it's not possible to supply wrapper flags to
the rule - only through the command line.

The only alternative is to create a sh_binary wrapper to the
wrapper wrapping the emulator wrapping our code invoked by
the bazel wrappers. In short this PR can avoid yet one more
wrapping.