gee: re-enable autostash (#780)
gee: re-enable autostash

The users are right: the "autostash" behavior is the right choice for
the "just function" philosophy of gee.  This PR re-enables autostash,
but adds a few checks for post-rebase merge conflict markers, just
in case a merge conflict was introduced when the stashed changes
got re-applied.

Tested: Used modified gee to run `gee rupdate` in a branch with uncommitted changes.