enkit outputs: Support persistent tunnels (#786)
When doing `enkit outputs mount`, the destination host resolves to an IP
in the network due to DNS spoofing to redirect traffic to a persistent
tunnel. The logic correctly recognizes that it does not need to create
an additional tunnel, but then attempts to connect to the remote port
(traefik on 8080, in this case) rather than the persistent tunnel port
(8001).

This change expands the tunnel decision logic to accommmodate the
"persistent tunnel on the gateway" case. When this scenario is detected,
no tunnel will be created, but the port will be replaced with
`tunnelPort` in the returned value. By setting the
`--tunnel-listen-port` flag, the target port can be changed to the port
the tunnel on the gateway is listening on.

Tested:
* `bazel build //guidelines/...`
* `enkit outputs mount -i 07dc68b2-56b6-4f9f-bb93-a88bb7bb5003 --loglevel-console=debug --tunnel-listen-port=8001`
  * Saw correct debug log message
* Confirm all files are readable: `find ~/outputs/07dc68b2-56b6-4f9f-bb93-a88bb7bb5003/ -type f | xargs md5sum`

Jira: INFRA-2343