remote: Use "target" config instead of "host" for remote_run rule
The reference below gets into more details, but the dependencies of a
remote_run rule are really runtime dependencies that need to be compiled
for the target architecture. We want the dependencies to be built as if
they were built as a top-level target on the remote machine. While tools
might follow the "exec" config, that does not apply here either.

Ref: https://bazel.build/extending/rules#configurations
Jira: SF-478

Signed-off-by: Raghu Raja <raghu@enfabrica.net>