gee pr_checks: capture the build results url (#788)
gee pr_checks: capture URL for buildbuddy

This change saves the user a little time when debugging presubmit failures.
`gee pr_checks` now reports a summary of the error messages from the
presubmit job, including the line that contains a URL that points to
the specific buildbuddy page that contains the information the user
probably needs to fix the issue.

Tested: Ran the updated pr_checks command in a branch with a build failure.