machinist: Flag off most machine configuration
This change adds a new flag that disables much of the static machine
configuration done, so that machinist does not fight with current/future
puppet configuration on startup.

This flag defaults to false for the Puppet case, and needs to be
manually enabled in instances where machine configuration is still
desired.

Dynamic steps (writing out keys, certs, etc. received from the auth
server at runtime) are not affected by this flag - only static
configuration and installation steps are controlled.

Tested: not yet

Jira: INFRA-2550