lib/oauth: support multiple verifiers.
Background:
At the end of the oauth process, once a token has been obtained,
a verifier is invoked. This verifier checks the claims of the token,
and creates an identity for the user.

Today, a single verifier can be specified. Generally, the verifier
makes a call to the provider API to fetch eg, the email or information
of the user, or uses a key/certificate to extract and verify it from
the token.

This PR is meant to introduce no functional changes, just juggles
the code around. Specifically:
- there is now a "Verifier" interface, that defines 1) the scopes
  the oauth token needs for the verifier to work 2) provides a
  function to actually verify the token.
- moves the code around in the ogoogle and ogithub libraries to
  implement the verifier interface. Note that the code is exactly
  the same, todo by todo and comment by comment,  just moved in
  the interface.
- changes factory.go to use multiple verifiers.

Also, adds some defense in depth so if verifiers succeed but return
no valid credentials, the authentication fails.

Tested:
Server started with dev credentials after the change successfully
authenticates.